:fire: :tractor: Continue cleaning up the handling of step settings
